requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
bcrypt
cryptography
gunicorn
python-dotenv
